Make PointPattern instances picklable
Remove all memoized closures in the pointpatterns module, since objects
which stores functions that are not declared on the module level can not
be pickled. Override the __reduce__ and __setstate__ methods for classes
inheriting from shapely classes. Make the cache name used by
memoize_method availble as a class member