Systematically inherit operations for IntensityMap
The place to inherit these from is numpy.ma.MaskedArray. Hence, the
standard internal representation of the intensities in IntensityMap is
changed from ndarrays (with possible nans) to masked arrays. This is
also beneficial in that masked arrays can have any dtype, while the
presence of nans requires float arrays.