Improve cell feature array
The cell feature array now consists of well thought out features
parametrizing shape and size independently, and giving a distance that is
independent of measurement units. Moreover, the method for computing the
feature array takes an optional argument for specifying the relative
weight of the grid scale compared to shape features.